refactor: transpose_column functions are moved to a seperate class
